passaggio da double a float definizione dei traits per il tipo di dato da mettere nelle foglie refactoring del metodo di intersezione per evitare il più possibile di creare dati in memoria che vengano subito dopo eliminati iniziale passaggio (veder de capirghe qualcossa) ai Pointer Container nella classe Octree: per il momento l'uso è solo per i ticket ma c'è da capire se può essere esteso

+++ source/common/EulerAngles.cpp
@@ -11,7 +11,7 @@ EulerAngles::EulerAngles() :
 	ALPHA(0), BETA(0), GAMMA(0) {
 }
 
-EulerAngles::EulerAngles(double alpha, double beta, double gamma) :
+EulerAngles::EulerAngles(float alpha, float beta, float gamma) :
 	ALPHA(alpha), BETA(beta), GAMMA(gamma) {
 }
