+++ source/common/Rototraslation.hpp
@@ -26,15 +26,15 @@ public:
 		return os;
 	}
 	
-	Eigen::Isometry3d asEigen() const {
+	Eigen::Isometry3f asEigen() const {
 		
 		/* I have to go this way in order do bypass an eigen error handling
 		 * rotation matrix: doing something like
 		 * Isometry3d = Translation3d * Matrix3d results in a wrong Isometry
 		 * matrix (both rotation and translation)
 		 */
-		Eigen::Isometry3d tmpRot(ROTATION.asEigen());
-		Eigen::Isometry3d rototras = TRASLATION.asTranslation() * tmpRot;
+		Eigen::Isometry3f tmpRot(ROTATION.asEigen());
+		Eigen::Isometry3f rototras = TRASLATION.asTranslation() * tmpRot;
 		
 		return rototras;
 	}
